Fixed race condition &  minor bug with the AWS URL
It appears that there was a race condition when creating
downloaded directories with the S2 AWS archive downloader.
This has now been fixed (not very elegantly, I must add), but
some non-useful code has been left in, in case archaeologists
want to delve in it in the future.

A second problem was that the AWS URL had been corrupted by
what can only be a thick finger problem. This has now been
solved. It now works for me.