MODIS L1 data downloader from LAADS
I added this convenience downloader. As it is hard to figure out
what granules you want over some area of interest, you need to
do an online query, get the results as a JSON file, and then this
script will download them.

One thing that is probably a good idea is something that creates
a directory structure based on the product name, but that's easy
to add.

Seems to work fine.