Better MODIS downloading
I realised that sometimes you get kicked out of the MODIS server,
and files don't finish downloading cleanly. A way around this is to
download files with a different name, and rename them once they're
done. The downloader checks the output folder, and if the files are
present in their final form, will not download them.